refactor: 요구사항에 맞게 추가적 엔티티 삭제
